controllers/export_controller.py
import os
from PIL import Image
from reportlab.pdfgen import canvas
from reportlab.lib.pagesizes import A4, A3, LETTER, LEGAL

class ExportController:
    def __init__(self, db):
        self.db = db
        self.page_sizes = {
            'A4': A4,
            'A3': A3,
            'LETTER': LETTER,
            'LEGAL': LEGAL
        }
    
    def export_component_as_image(self, component, output_path: str, format: str = 'PNG'):
        """Export a single component as an image"""
        # Create image from component properties
        width = component.properties.get('dimensions', {}).get('width', 100)
        height = component.properties.get('dimensions', {}).get('height', 100)
        
        # Create new image
        img = Image.new('RGB', (width, height), 'white')
        
        # Add component content
        self._render_component_to_image(component, img)
        
        # Save image
        img.save(output_path, format)
        return output_path
    
    def export_components_as_pdf(self, components: list, output_path: str, page_size: str = 'A4'):
        """Export multiple components as a PDF"""
        # Create PDF
        c = canvas.Canvas(output_path, pagesize=self.page_sizes.get(page_size, A4))
        
        # Add components to PDF
        for component in components:
            self._add_component_to_pdf(c, component)
            c.showPage()
        
        c.save()
        return output_path
    
    def _render_component_to_image(self, component, img):
        """Render component content onto an image"""
        # Implementation depends on component type and properties
        pass
    
    def _add_component_to_pdf(self, canvas, component):
        """Add a component to a PDF canvas"""
        # Implementation depends on component type and properties
        pass 