requirements.txt
customtkinter>=5.2.0
Pillow>=10.0.0
reportlab>=4.0.0
pandas>=2.0.0
openpyxl>=3.1.0  # For Excel file support 
qrcode>=7.4.2